pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "after-market-agent"
version = "0.1.0"
description = "Backend scaffold for the After Market Agent service."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10,<4.0",
    "alembic>=1.13,<2.0",
    "fastapi>=0.115,<1.0",
    "pydantic>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "python-dotenv>=1.0,<2.0",
    "sqlalchemy>=2.0,<3.0",
    "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27,<1.0",
    "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["backend/tests"]
